refactor: Drop withMobxStore indirection – move logic into _app.js
This makes it easier to see how the pieces fit together and moves
_app.js closer to the example given by https://nextjs.org/docs/#custom-app
